==> accident_rate_Madrid/1_Access_binary_format_files.ipynb <==
# Access through files: binary formats

The following is a simple example of data access using Python files, using the facilities provided by the **Pandas** library.

The input file corresponds to traffic accidents in the city of Madrid in 2019 and can be downloaded from the following page.

import pandas as pd

accidentes = pd.read_csv('./2019_Accidentalidad.csv', 
                        delimiter = ";",
                        encoding = 'utf-8',
                        parse_dates = ["fecha"],
                        na_values = ['-', 'Desconocido'])

accidentes[:5]

columnas_dropped = ['cod_distrito', 'lesividad', 'coordenada_x_utm', 
                    'coordenada_y_utm', 'positiva_alcohol', 'positiva_droga']

accidentes.drop(columnas_dropped, axis = 1, inplace = True)

accidentes[:5]

accidentes.to_parquet('./accidentes.parquet')
