==> accident_rate_Madrid/2_Format_standardization.ipynb <==
# Format standardization

The following code shows an example of data standardization to facilitate its exploitation, based on the same data used in the example of file *1_Access_binary_format_files*.

In this case, the same page where the data were downloaded offers a document describing how to interpret the data provided: in this case, it is very convenient to translate all this contextual information into explicit information in the data, as will be seen in the case of the new “Severity” field or in the limitation of the possible values for those fields that have them defined as such.

import pandas as pd
import numpy as np

1. We read the previously saved data, having discarded certain columns in order to match the file downloaded in 2020, as it had undergone some modifications.

accidentes = pd.read_parquet('./accidentes.parquet')

accidentes.shape

2. We generated a new field to code the severity levels according to “lesividad”, as specified in the data description document.

accidentes.info()

accidentes.cod_lesividad.isnull().sum()

c_gravedad = pd.api.types.CategoricalDtype(categories = ['Ileso', 'Leve', 'Grave','Fallecido'], ordered = True)

dict_gravedad = {14.0: 'Ileso', 3.0: 'Grave', 4.0: 'Fallecido'}

accidentes['gravedad'] = accidentes['cod_lesividad'].apply(
lambda x: dict_gravedad.get(x, 'Leve') if ~np.isnan(x) else 'Ileso').astype(c_gravedad)

3. We decode the time and add it to the date to have a date-time type field, easier to use.

# Convert the column 'hora' to a datetime object
accidentes['hora'] = pd.to_datetime(accidentes['hora'], format = '%H:%M:%S').dt.strftime('%H:%M:%S')

# Combine the column 'fecha' and 'hora'.
accidentes['fecha'] += pd.to_timedelta(accidentes['hora'])

# We eliminate columns that are no longer needed
accidentes = accidentes.drop(columns = ['hora', 'localizacion', 'numero'])

4. We convert all columns starting from the third one into categorical types.

accidentes.iloc[:,2:] = accidentes.iloc[:,2:].astype('category')

accidentes[:5]

accidentes.to_parquet('./accidentes1.parquet')

accidentes['gravedad'].value_counts()

The penultimate lines of the code require a little explanation.

- The data interpretation document gives a definition of the different possible values of the “LESIVIDAD*” field, and adds that value 3 is considered “grave”, value 14 is considered “uninjured”, value 4 corresponds to the death of the injured person, and all other values are considered “leve”.

- In another part of the document, it is specified that if the injury value is not present, it is because there was no victim in the accident requiring medical attention (which we are going to consider here as “Ileso”, although it could be that in some cases it was “Leve”).

- All the gibberish in that document is translated into three lines of code, and once the new “GRAVEDAD” field is created (which does not add any information that was not present in the original data, of course) it is not necessary to consider it again during the rest of the analyses made with that data.

- In addition, the order of the injury codes is arbitrary, while the severity codes are ordered, making it possible to compare them and do analyses such as “in what percentage of accidents between a car and a motorcycle the severity of the motorcycle driver is greater than that of any of the occupants of the car”, for example.

This is an example of a highly recommended practice in data analysis: translating implicit codings into explicit ones, and translating confusing codings into codings that are easy to understand and use.
